+++ packages/terra-site/src/examples/form-radio/LongTextRadioExample.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Radio from 'terra-form-radio/lib/Radio';
+
+const hiddenLabelRadioExample = () => (
+  <Radio
+    id="long-text"
+    name="longText"
+    labelText={'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua ' +
+              'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua'}
+  />);
+
+export default hiddenLabelRadioExample;